[INTERNAL] Rename alignment directory to alignment_file.
