+++ include/seqan3/io/alignment_file/detail.hpp
